backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.44
matplotlib==3.9.2
mplfinance==0.12.10b0
Pillow==10.4.0
requests==2.32.3
